1 - Anyway grabbed from the openjfx the source files (GNU licensed) needed to get observable lists working here without needing the entire javafx library. 2 - Updated mockito since it caused some reflection error on java 9+ 3 - optimized imports for test set

+++ src/test/java/uk/ac/ox/oxfish/fisher/strategies/destination/YearlyIterativeDestinationStrategyTest.java
@@ -66,7 +66,7 @@ public class YearlyIterativeDestinationStrategyTest {
 
 
         //cashflow is x+y
-        doAnswer(invocation -> delegate.getFavoriteSpot().getGridX() + delegate.getFavoriteSpot().getGridY()).when(
+        doAnswer(invocation -> (double)(delegate.getFavoriteSpot().getGridX() + delegate.getFavoriteSpot().getGridY())).when(
                 fisher).getBankBalance();
         when(fisher.balanceXDaysAgo(360)).thenReturn(0d);
